Added const array of syscall args registers
In order to allow the tool check whether any of the syscall arguments
comes from an uninitialized read, we defined a constant array containing
the list of registers that hold syscall arguments.

Since they are different according to the architecture (X86 vs X86-64)
we defined them in Platform.h, after we detected the architecture being
used.