Fix merge_reports in case of overlaps in mmap heap
In case an uninitialized read happened on a chunk allocated in a mmapped
heap, the base addresses were not added to the list of allocated
libraries and pages, and the script assumed it was on the main heap, if
it was not a stack address.

The script has been fixed,so if an uninitialized read happens on a
mmapped heap, we can see the base address, we can distinguish it from
the "main heap" (if any) and the merge_report script won't fail finding
the base address in the parsed data.

In order to make it clear that these heaps are not the "main" heap (e.g. the
one managed through brk/sbrk on Linux), they are called "Heap N", where
N is the order with which the program allocated them, starting by 1.