Fixed direct memory copy of indirect uninit read
If an uninitialized read is an indirect one (reading uninitialized bytes
stored into memory by previous instructions) and it is a direct memory
copy instruction (e.g. push instruction with a memory operand), does not
consider it a usage of uninitialized bytes, but copy them into the new
memory location.