Fixed argvfuzz and memTracer.py argv getter
In order to avoid ignoring some of the initial arguments passed to the
program, memTracer.py does return the whole argument array, without
removing the first element.

However, this must be true also in the argv fuzzing library.
In order to fix its behavior, the executable path is set as a first
thing at index 0 of the returned array, instead of modifying the first
element after it has been returned.
Of course, this means that the first argument that is retrieved from the
input file is stored at index 1 of the array.