Fixed segfault caused by mremap reallocs
If a malloc allocate a single chunk using mmap, calling realloc on that
chunk will use mremap to resize or move the allocated chunk.
That was not detected by MemTrace, thus causing a segmentation fault as
soon as the new chunk was freed.

Also, fixed possible segfault sources in rare cases where a big chunk is
allocated without using mmap or mremap.