Fixed bugs potentially causing false positives
In some cases, the following bugs created some false positives.
They were limited in number, but they might confuse the user.

[*] Tag set in high byte registers:
If the src register does not have a high byte register, but the dst
register does have a high byte register, getting the corresponding
registers of the src one didn't get the high byte register of the dst.
So, the tag set of the dst high byte register was not updated, and
therefore it would have been considered as a pure uninitialized read
instead of an indirect uninitialized read.

[*] Temporary write storage clearing on Realloc
A call to Realloc might free a buffer and allocate a new one.
Both free and malloc handlers in MemTrace use the same structure as a
temporary storage for writes, so that the tool can report the write
accesses performed during free/malloc.
In both the cases, however, the temporary storage is cleared when the
handler finishes.
By doing this after the call to FreeAfter called by MallocAfter due to
the execution of realloc, FreeAfter might delete some writes added
during the copy of bytes performed by realloc.
This might have generated false uninitialized reads.