Handled brk used to release memory
If a call to *free* or *malloc_trim* release memory to the OS, the tool
will remove all the writes accessing the released memory.
This way, if the same memory portion is used again, the tool will not
erroneously report write accesses whose byte have been deleted by the
release of memory.

Notice that this kind of operation is quite expensive, but it happens
very rarely (only found once testing with binaries from ubuntu server
manifest)