Fixed StackShadow::reset in case of 32 bit binary
If MemTrace is executed on a 32 bits architecture, and therefore the
stack pointer is 4 bytes long and aligned to 4 bytes boundaries on
return, the reset method will check whether the passed address is
aligned to an 8 bytes boundary.

If it is, we need to reset only the 4 LSB of the current shadow address,
and then all the shadow address above it, leaving the 4 MSB of the
current shadow address untouched.

Example:
if on x86 addr is 0x7fff00, we must reset bytes in the range
0x7fff00 ~ 0x7fff03, while bytes in the range
0x7fff4 ~ 0x7fff07 are not reset, as it is a stack cell higher than the
address passed to reset.